fix: rebase 이후 에러 수정 및 중복 파일 정리
