+++ src/domains/exercise-logs/repository/index.js
@@ -8,26 +8,26 @@ class ExerciseLogRepository {
   async create(data) {
     return await this.prisma.exerciseLog.create({ data });
   }
-}
 
   //Get
-  findLatestByJournalId=async(journalId)=>{
+  findLatestByJournalId = async (journalId) => {
     return await this.prisma.exerciseLog.findFirst({
-      where: {journalId},
-      orderBy: {endTime: 'desc'}
+      where: { journalId },
+      orderBy: { endTime: 'desc' },
     });
-  }
+  };
 
   //지금까지 얻은 포인트의 총합. 한 journalId에 저장된 earnedPoints를 모두 더함.
   //아직 운동한적 없어서 기록이 없으면 0
-  sumEarnedPoints=async(journalId)=>{
+  sumEarnedPoints = async (journalId) => {
     const result = await this.prisma.exerciseLog.aggregate({
-      where: {journalId},
+      where: { journalId },
       _sum: {
-        earnedPoints: true
-      }
+        earnedPoints: true,
+      },
     });
     return result._sum.earnedPoints || 0;
-  }
+  };
+}
 
 export default ExerciseLogRepository;
