--- src/domains/journals/controller/journalsController.js
@@ -1,7 +1,0 @@
-export class JournalsController {
-    constructor(/* journalsService */) {
-        // this.journalsService = journalsService; // JournalsService 인스턴스 주입
-    }
-}
-
-export const journalsController = new JournalsController();
