--- src/domains/journals/repository/journalsRepositury.js
@@ -1,7 +1,0 @@
-export class JournalsRepository {
-    constructor(prisma) { 
-        this.prisma = prisma || new PrismaClient(); 
-    }
-}
-
-export const journalsRepository = new JournalsRepository();
